style: enforce strict typing and ruff formatting
Apply ruff/black formatting, add explicit type hints for arrays, and wrap long lines in core and optimization modules.

+++ scripts/laminar_logprob_probe.py
@@ -104,7 +104,7 @@ def build_init_dict(
 
     def to_z(name: str, value: float) -> float:
         s = scalings[name]
-        return (value - s.center) / s.scale
+        return float((value - s.center) / s.scale)
 
     physical = {
         "contrast_0": contrast,
